팀별 게시판 작성, 수정, 삭제 기능 추가

+++ team2/src/main/java/com/team2/board/controller/ReplyController.java
@@ -1,5 +1,7 @@
 package com.team2.board.controller;
 
+import java.security.Principal;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,6 +13,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.team2.board.model.ReplyVO;
 import com.team2.board.service.IBoardService;
+import com.team2.member.model.Member;
+import com.team2.member.service.IMemberService;
 
 @RestController
 @RequestMapping("/reply")
@@ -18,17 +22,29 @@ public class ReplyController {
 	
 	@Autowired
 	IBoardService boardService;
+	@Autowired
+	IMemberService memberService;
 	
 	// 댓글 작성 
 	@PostMapping(value="/create/{teamId}") 
-	public ReplyVO createReply(@RequestBody ReplyVO reply) {
+	public ReplyVO createReply(@RequestBody ReplyVO reply, 
+			@PathVariable String teamId, Principal principal) {
+		Member member = memberService.selectMember(principal.getName());
+		if(!teamId.equals(Integer.toString(member.getTeamId()))) {
+			return null; 
+		}
 		boardService.createReply(reply, boardService.maxReplyId());
 		return reply;
 	}
 	
 	// 댓글 삭제 
 	@DeleteMapping("/delete/{teamId}/{replyId}")
-	public String deleteReply(@PathVariable int replyId) {
+	public String deleteReply(@PathVariable int replyId, 
+			@PathVariable String teamId, Principal principal) {
+		Member member = memberService.selectMember(principal.getName());
+		if(!teamId.equals(Integer.toString(member.getTeamId()))) {
+			return "fail"; 
+		}
 		try {
 			boardService.deleteReply(replyId);
 		} catch (Exception e) {
@@ -40,7 +56,12 @@ public class ReplyController {
 	
 	// 댓글 수정 
 	@PutMapping("/update/{teamId}")
-	public ReplyVO updateReply(@RequestBody ReplyVO reply) {
+	public ReplyVO updateReply(@RequestBody ReplyVO reply,
+			@PathVariable String teamId, Principal principal) {
+		Member member = memberService.selectMember(principal.getName());
+		if(!teamId.equals(Integer.toString(member.getTeamId()))) {
+			return null; 
+		}
 		boardService.updateReply(reply);
 		return reply;
 	}
